+++ src/aks-preview/HISTORY.md
@@ -6,6 +6,7 @@ Release History
 0.5.26
 +++++
 * Add podMaxPids argument for node config
+* Correct containerLogMaxSizeMb to containerLogMaxSizeMB in customized kubelet config
 
 0.5.25
 +++++